+++ frontend/src/components/ui/compound/NavigationBar/Menu.tsx
@@ -1,22 +1,12 @@
 import { AdjustmentsHorizontalIcon, Bars3Icon, UserCircleIcon } from '@heroicons/react/24/outline';
-import { ArrowRightOnRectangleIcon, PencilSquareIcon } from '@heroicons/react/24/solid';
 import { css } from 'styled-components';
 
-import type { PropsWithChildren } from 'react';
 import { HiArrowPath, HiOutlineChatBubbleOvalLeftEllipsis } from 'react-icons/hi2';
 
-import { useQueryClient } from '@tanstack/react-query';
-
 import { useExternalValue } from '@utils/external-state';
-import { logout } from '@utils/login';
 
 import { modalActions } from '@stores/layout/modalStore';
 import { memberTokenStore } from '@stores/login/memberTokenStore';
-import {
-  selectedCapacitiesFilterStore,
-  selectedCompaniesFilterStore,
-  selectedConnectorTypesFilterStore,
-} from '@stores/station-filters/serverStationFiltersStore';
 
 import Button from '@common/Button';
 import FlexBox from '@common/FlexBox';
@@ -25,57 +15,23 @@ import ServerStationFilters from '@ui/ServerStationFilters';
 import StationListWindow from '@ui/StationList/StationListWindow';
 import StationSearchWindow from '@ui/StationSearchWindow';
 import LogoIcon from '@ui/Svg/LogoIcon';
-import CarModal from '@ui/modal/CarModal/CarModal';
+import PersonalMenu from '@ui/compound/NavigationBar/PersonalMenu';
 import LoginModal from '@ui/modal/LoginModal/LoginModal';
 
 import { displayNoneInMobile, displayNoneInWeb } from '@style/mediaQuery';
 
 import { EMPTY_MEMBER_TOKEN, MOBILE_BREAKPOINT } from '@constants';
-import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
 
-import PopupMenu from '../PopupMenu';
 import { useNavigationBar } from './hooks/useNavigationBar';
 
 const Menu = () => {
   const { openBasePanel } = useNavigationBar();
 
   const memberToken = useExternalValue(memberTokenStore);
-  const queryClient = useQueryClient();
-
-  const { openModal } = modalActions;
 
   const handleClickLoginIcon = () => {
     modalActions.openModal(<LoginModal />);
   };
-
-  const loginMenus: PropsWithChildren<{ onClick: () => void }>[] = [
-    {
-      children: (
-        <>
-          <PencilSquareIcon width="1.8rem" color="#333" /> 차량등록
-        </>
-      ),
-      onClick: () => {
-        openModal(<CarModal />);
-      },
-    },
-    {
-      children: (
-        <>
-          <ArrowRightOnRectangleIcon width="1.8rem" color="#333" /> 로그아웃
-        </>
-      ),
-      onClick: () => {
-        logout();
-
-        selectedCapacitiesFilterStore.setState(new Set([]));
-        selectedConnectorTypesFilterStore.setState(new Set([]));
-        selectedCompaniesFilterStore.setState(new Set([]));
-
-        queryClient.invalidateQueries({ queryKey: [QUERY_KEY_STATION_MARKERS] });
-      },
-    },
-  ];
 
   return (
     <FlexBox css={[fixedPositionCss, paddingCss, borderCss, flexCss]} noRadius="all" nowrap>
@@ -98,7 +54,7 @@ const Menu = () => {
         <AdjustmentsHorizontalIcon width="2.8rem" stroke="#333" />
       </Button>
       {memberToken !== EMPTY_MEMBER_TOKEN ? (
-        <PopupMenu menus={loginMenus} />
+        <PersonalMenu />
       ) : (
         <Button aria-label="로그인 하기" onClick={handleClickLoginIcon}>
           <UserCircleIcon width="2.8rem" stroke="#333" />
